Refactor WindowManager to a register-show-run workflow
Previously, to show a window in the window manager, one would call `Run`
to start the application, and `Show` to change windows once it had
started.

This made it difficult for the window manager to reason about windows,
becuase it would only ever know about the window currently being shown.

To improve matters, the workflow is now
1. Call `RegisterWindow` with a string name to tell the window manager
about the window
2. Call `ShowWindow` with a string name to make a window visible
3. Call `Run` to start the application. The first window to be shown
will be the current window, as determined by the last `ShowWindow` call.