Introduce a message bus
Lots of this application is event driven. Currently, it's quite tightly
bound. To make it easier to test and maintain, we need to make it more
loosely coupled.

To get there, we introduce a message bus.

Event producers (such as checking for a new version, or new messages
coming in from Discord) can publish events independently of event
consumers (such as UI components). We're able to test the two items
independently by testing how they respond to published messages, or
otherwise checking they publish a message at the right now.