Implement version updates UI as a true dialog
The window manager has always outlined dialog functionality, but never
implemented it. This commit implements it.

Showing a dialog by calling `WindowManager.Dialog` now draws the dialog
by calling its `Show` method, and activates it by calling its `Open`
method. Once the dialog invokes the closer function paased to it, the
window manager reinstates the previous window.

Checking for version updates is now implemented as a dialog. The
checking for the version and the control of the UI are composed in the
app composition root.

The UI and the version logic are connected over the event bus.